refactor(qwen35): extract DDTree into src/qwen35/ddtree.{h,cpp}
Move DDTree struct, build_ddtree(), follow_verified_tree(), and
extract_draft_topk() from test_dflash.cpp (inline statics) into a
proper library module. test_dflash.cpp now includes ddtree.h and uses
the library symbols via using-declarations.

This is part of Phase 3: extracting qwen35 infrastructure from test/
into src/ for reuse by a future Qwen35Backend.
